requirements.txt
django
django-environ
boto3
django-storages
psycopg2-binary
botocore
crispy-bootstrap5
django-crispy-forms
django-storages
django-cors-headers
django-utils-six
django-tinymce
django-allauth
pytz
django-extensions
Werkzeug
pyOpenSSL
pycryptodome
